Update the datatype of tables using INTEGER
- Includes fixes in imports for new jobs
- Includes fix in column name where engraftment material is taken from

+++ scripts/init.sql
@@ -126,19 +126,19 @@ ALTER TABLE patient_sample
     REFERENCES model (id);
 
 CREATE TABLE engraftment_site (
-    id INTEGER NOT NULL,
+    id BIGINT NOT NULL,
     name TEXT NOT NULL,
     PRIMARY KEY (id)
 );
 
 CREATE TABLE engraftment_type (
-    id INTEGER NOT NULL,
+    id BIGINT NOT NULL,
     name TEXT NOT NULL,
     PRIMARY KEY (id)
 );
 
 CREATE TABLE engraftment_material (
-    id INTEGER NOT NULL,
+    id BIGINT NOT NULL,
     name TEXT NOT NULL,
     PRIMARY KEY (id)
 );
